third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-autospace-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-autospace value 'initial'
  assert_equals: expected "normal" but got "no-autospace"
[FAIL] Property text-autospace value 'auto'
  assert_true: 'auto' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha'
  assert_true: 'ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-numeric'
  assert_true: 'ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha ideograph-numeric'
  assert_true: 'ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation'
  assert_true: 'punctuation' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation ideograph-alpha'
  assert_true: 'punctuation ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation ideograph-alpha ideograph-numeric'
  assert_true: 'punctuation ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha insert'
  assert_true: 'ideograph-alpha insert' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-numeric insert'
  assert_true: 'ideograph-numeric insert' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha ideograph-numeric insert'
  assert_true: 'ideograph-alpha ideograph-numeric insert' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation insert'
  assert_true: 'punctuation insert' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation ideograph-alpha insert'
  assert_true: 'punctuation ideograph-alpha insert' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert ideograph-alpha'
  assert_true: 'insert ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert ideograph-numeric'
  assert_true: 'insert ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert ideograph-alpha ideograph-numeric'
  assert_true: 'insert ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert punctuation'
  assert_true: 'insert punctuation' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert punctuation ideograph-alpha'
  assert_true: 'insert punctuation ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'insert punctuation ideograph-alpha ideograph-numeric'
  assert_true: 'insert punctuation ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha replace'
  assert_true: 'ideograph-alpha replace' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-numeric replace'
  assert_true: 'ideograph-numeric replace' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'ideograph-alpha ideograph-numeric replace'
  assert_true: 'ideograph-alpha ideograph-numeric replace' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation replace'
  assert_true: 'punctuation replace' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'punctuation ideograph-alpha replace'
  assert_true: 'punctuation ideograph-alpha replace' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace ideograph-alpha'
  assert_true: 'replace ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace ideograph-numeric'
  assert_true: 'replace ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace ideograph-alpha ideograph-numeric'
  assert_true: 'replace ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace punctuation'
  assert_true: 'replace punctuation' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace punctuation ideograph-alpha'
  assert_true: 'replace punctuation ideograph-alpha' is a supported value for text-autospace. expected true got false
[FAIL] Property text-autospace value 'replace punctuation ideograph-alpha ideograph-numeric'
  assert_true: 'replace punctuation ideograph-alpha ideograph-numeric' is a supported value for text-autospace. expected true got false
Harness: the test ran to completion.
